[camera][next] split delegates, fix access levels
