+++ packages/expo-camera/ios/Next/BarCodeScanner.swift
@@ -9,20 +9,20 @@ class BarCodeScanner: NSObject {
 
   // MARK: - Properties
 
-  private var session: AVCaptureSession
-  private var sessionQueue: DispatchQueue
-  private var metadataOutput: AVCaptureMetadataOutput?
-  private var videoDataOutput: AVCaptureVideoDataOutput?
-
-  private var settings = BarCodeScannerUtils.getDefaultSettings()
-  private var zxingBarcodeReaders: [AVMetadataObject.ObjectType: ZXReader] = [
+  private let session: AVCaptureSession
+  private let sessionQueue: DispatchQueue
+  private let zxingCaptureQueue = DispatchQueue(label: "com.zxing.captureQueue")
+  
+  internal var metadataOutput: AVCaptureMetadataOutput?
+  internal var videoDataOutput: AVCaptureVideoDataOutput?
+  internal var settings = BarCodeScannerUtils.getDefaultSettings()
+  internal var zxingBarcodeReaders: [AVMetadataObject.ObjectType: ZXReader] = [
     AVMetadataObject.ObjectType.pdf417: ZXPDF417Reader(),
     AVMetadataObject.ObjectType.code39: ZXCode39Reader()
   ]
 
-  private var zxingFPSProcessed = 6.0
-  private var zxingCaptureQueue = DispatchQueue(label: "com.zxing.captureQueue")
-  private var zxingEnabled = true
+  internal var zxingFPSProcessed = 6.0
+  internal var zxingEnabled = true
 
   init(session: AVCaptureSession, sessionQueue: DispatchQueue) {
     self.session = session
@@ -81,20 +81,16 @@ class BarCodeScanner: NSObject {
     }
 
     if metadataOutput == nil || videoDataOutput == nil {
-      session.beginConfiguration()
       addOutputs()
-      session.commitConfiguration()
-
       if metadataOutput == nil {
         return
       }
     }
 
     var availableRequestedObjectTypes: [AVMetadataObject.ObjectType] = []
-    let requestedObjectTypes = settings[BARCODE_TYPES_KEY] ?? []
     let availableObjectTypes: [AVMetadataObject.ObjectType] = metadataOutput?.availableMetadataObjectTypes ?? []
 
-    for type in requestedObjectTypes where availableObjectTypes.contains(type) {
+    for type in settings[BARCODE_TYPES_KEY] ?? [] where availableObjectTypes.contains(type) {
       availableRequestedObjectTypes.append(type)
     }
 
@@ -102,10 +98,7 @@ class BarCodeScanner: NSObject {
   }
 
   func stopBarCodeScanning() {
-    session.beginConfiguration()
     removeOutputs()
-    session.commitConfiguration()
-
     if isScanningBarcodes {
       onBarCodeScanned?(nil)
     }
@@ -142,6 +135,8 @@ class BarCodeScanner: NSObject {
   }
 
   private func addOutputs() {
+    session.beginConfiguration()
+
     if metadataOutput == nil {
       let output = AVCaptureMetadataOutput()
       output.setMetadataObjectsDelegate(self, queue: sessionQueue)
@@ -161,9 +156,13 @@ class BarCodeScanner: NSObject {
         videoDataOutput = output
       }
     }
+    
+    session.commitConfiguration()
   }
 
   private func removeOutputs() {
+    session.beginConfiguration()
+  
     if let metadataOutput {
       if session.outputs.contains(metadataOutput) {
         session.removeOutput(metadataOutput)
@@ -177,57 +176,7 @@ class BarCodeScanner: NSObject {
         self.videoDataOutput = nil
       }
     }
-  }
-}
-
-// MARK: - Delegates
-
-extension BarCodeScanner: AVCaptureMetadataOutputObjectsDelegate, AVCaptureVideoDataOutputSampleBufferDelegate {
-  func metadataOutput(_ output: AVCaptureMetadataOutput, didOutput metadataObjects: [AVMetadataObject], from connection: AVCaptureConnection) {
-    guard let settings = settings[BARCODE_TYPES_KEY], let metadataOutput else {
-      return
-    }
-
-    for metadata in metadataObjects {
-      let codeMetadata = metadata as? AVMetadataMachineReadableCodeObject
-      for barcodeType in settings {
-        if zxingBarcodeReaders[barcodeType] != nil {
-          continue
-        }
-
-        if let codeMetadata {
-          if codeMetadata.stringValue != nil && codeMetadata.type == barcodeType {
-            onBarCodeScanned?(BarCodeScannerUtils.avMetadataCodeObjectToDictionary(codeMetadata))
-          }
-        }
-      }
-    }
-  }
-
-  func captureOutput(_ output: AVCaptureOutput, didOutput sampleBuffer: CMSampleBuffer, from connection: AVCaptureConnection) {
-    guard let barCodeTypes = settings[BARCODE_TYPES_KEY],
-      let metadataOutput,
-      zxingEnabled else {
-      return
-    }
-
-    let kMinMargin = 1.0 / zxingFPSProcessed
-    let presentTimeStamp = CMSampleBufferGetPresentationTimeStamp(sampleBuffer)
-
-    var curFrameTimeStamp = 0.0
-    var lastFrameTimeStamp = 0.0
-
-    curFrameTimeStamp = Double(presentTimeStamp.value) / Double(presentTimeStamp.timescale)
-
-    if curFrameTimeStamp - lastFrameTimeStamp > Double(kMinMargin) {
-      lastFrameTimeStamp = curFrameTimeStamp
-
-      if let videoFrame = CMSampleBufferGetImageBuffer(sampleBuffer),
-       let videoFrameImage = ZXCGImageLuminanceSource.createImage(from: videoFrame) {
-        self.scanBarcodes(from: videoFrameImage) { barCodeScannerResult in
-          self.onBarCodeScanned?(BarCodeScannerUtils.zxResultToDictionary(barCodeScannerResult))
-        }
-      }
-    }
+    
+    session.commitConfiguration()
   }
 }
